backend/pyproject.toml

[project]
name = "realdiag-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["fastapi", "uvicorn", "pyyaml"]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
